feat: interface의 concrete 클래스 @inject 처리 및 테스트 추가
