+++ nextstep-di/src/test/java/nextstep/di/factory/BeanScannerTest.java
@@ -2,7 +2,7 @@ package nextstep.di.factory;
 
 import nextstep.di.factory.example.JdbcQuestionRepository;
 import nextstep.di.factory.example.JdbcUserRepository;
-import nextstep.di.factory.example.MyQnaService;
+import nextstep.di.factory.example.MyQnaServiceImpl;
 import nextstep.di.factory.example.QnaController;
 import org.junit.jupiter.api.Test;
 
@@ -15,7 +15,7 @@ class BeanScannerTest {
     @Test
     void scan_Test() {
         BeanScanner beanScanner = new BeanScanner("nextstep.di.factory.example");
-        Set<Class<?>> actual = Set.of(QnaController.class, JdbcUserRepository.class, JdbcQuestionRepository.class, MyQnaService.class);
+        Set<Class<?>> actual = Set.of(QnaController.class, JdbcUserRepository.class, JdbcQuestionRepository.class, MyQnaServiceImpl.class);
         Set<Class<?>> expected = beanScanner.getBeans();
 
         assertThat(expected).containsAll(actual);
