+++ nextstep-di/src/test/java/nextstep/di/factory/example/MyQnaService.java
@@ -1,24 +1,7 @@
 package nextstep.di.factory.example;
 
-import nextstep.annotation.Inject;
-import nextstep.stereotype.Service;
+public interface MyQnaService {
+    UserRepository getUserRepository();
 
-@Service
-public class MyQnaService {
-    private UserRepository userRepository;
-    private QuestionRepository questionRepository;
-
-    @Inject
-    public MyQnaService(UserRepository userRepository, QuestionRepository questionRepository) {
-        this.userRepository = userRepository;
-        this.questionRepository = questionRepository;
-    }
-
-    public UserRepository getUserRepository() {
-        return userRepository;
-    }
-
-    public QuestionRepository getQuestionRepository() {
-        return questionRepository;
-    }
+    QuestionRepository getQuestionRepository();
 }
