+++ nextstep-di/src/test/java/nextstep/di/factory/example/MyQnaServiceImpl.java
@@ -1,0 +1,24 @@
+package nextstep.di.factory.example;
+
+import nextstep.annotation.Inject;
+import nextstep.stereotype.Service;
+
+@Service
+public class MyQnaServiceImpl implements MyQnaService {
+    private UserRepository userRepository;
+    private QuestionRepository questionRepository;
+
+    @Inject
+    public MyQnaServiceImpl(UserRepository userRepository, QuestionRepository questionRepository) {
+        this.userRepository = userRepository;
+        this.questionRepository = questionRepository;
+    }
+
+    public UserRepository getUserRepository() {
+        return userRepository;
+    }
+
+    public QuestionRepository getQuestionRepository() {
+        return questionRepository;
+    }
+}
